acme.sh: add missing variable init
If certificate issuance failed, the output directory is checked for
whether a previously-moved staging directory should be restored. But the
`staging_moved` variable used to store this was uninitialized and the
script uses `set -u`. So the script would error out without clearing the
created directory, causing further issues on subsequent runs.

Signed-off-by: Aditya Bhargava <rightaditya@gmail.com>